finished

+++ srcs/utils/utils_float/ft_round.c
@@ -6,7 +6,7 @@
 /*   By: mrodrigu <mrodrigu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/07 05:48:56 by mrodrigu          #+#    #+#             */
-/*   Updated: 2018/02/17 20:20:45 by mrodrigu         ###   ########.fr       */
+/*   Updated: 2018/02/19 11:02:56 by jagarcia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,12 +32,7 @@ char		*ft_round(char *str, int pos)
 {
 	char	ac;
 
-	//pos = (ft_strchr(str, 'e') - str - 1) +
-	//	ft_atoi(ft_strchr(str, 'e') + 1) + pos;
-	//ft_putnbr(pos);
-	//ft_putstr("\n");
-	//pos = ft_isdigit(str[pos + 1]) ? pos : pos + 1;
-	if (!ft_isdigit(str[pos]) || pos > ft_strlen(str) - 3)
+	if (!ft_isdigit(str[pos]) || pos > (int)ft_strlen(str) - 3)
 		return (str);
 	ac = 0;
 	while ((pos + 1) > 0)
